requirements.txt
matplotlib==3.5.1
numpy==1.23.5
pandas==1.4.2
pandas_market_calendars==4.1.4
pykalman==0.9.5
scikit_learn==1.0.2
statsmodels==0.13.2
tabula_py==2.7.0
tabulate==0.8.9
yfinance==0.2.24
